+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fix selected row style
 * Include current row with currentSelectedRows list in case of grid selection mode is row.
 * add missing arguments to gridConfiguration.copyWith
+* Update column menu example allow merging with the default menu
+
 
 ## [1.6.2] - 2025. 03. 19